breaks intro component up into smaller components. abstracts button logic
